Append and fix -m and -m2 switches documentation
S726-010 SA04-011

The console help screen about -m and -m2 switches was not quite right.
-m description "Minimum Ada recompilation, timestamp-based" was wrong
because it is not "timestamp-based", it is recompiling if checksums are
different even if the timestamps are the same. Correct description is
"Minimal recompilation of Ada", as before the "-m2" switch appearence,
because the main goal is to minimize recompilation.
-m2 description "Minimum Ada recompilation, checksum-based" is not right
because this switch is not about to avoid extra recompilation, but about
recompilation necessity on the same timestamp, when checksums are different.

Change-Id: Iee1dd46e7e5028508c2d6877db18d083ae62a2db