Lift assumption in gprbind of single digit gcc version
The assumption that the gcc major version number is the single digit
following "gcc version" in the version string emitted by gcc -v (such
as, say, 7 in "gcc version 7.5.0"), doesn't hold any more with the
move to gcc 10.

This change fixes this by fetching the first uninterrupted sequence
of digits in the string instead, then converting it to Natural.

	* gprbind.adb (Gprbind): Handle GCC_Version as a Natural
	instead of a Character. Add a Head_Natural_In local subprogram
	to extract the major version value from the version string.

Fix for T619-052

Change-Id: I4a4e2993f9ce4d2e4892273c2456647fa6b8f293