Fix target fallback when referencing implicit target
When a project sets a target by referencing the target
of imported project and that imported project does not specify
a target explicitly target fallback is not turned on since
there was no way to distinguish explicit target declaration
and reference of implicit target.

Add a new flag to Variable_Value that is set only when
evaluating 'Target attribute reference and there is no explicit
target declaration anywhere.
Check for that flag when deciding of target fallback
applicability.
Overwrite value of Target attribute in such case during
Apply_Config_File.

for T706-011

Change-Id: I8bc884c2d42efe1298772a0b6c73a8cebd5214a6