Parameter --src-subdirs together with Source_Files attribute
T902-036

All files in --src-subdirs directory should be available, even if it is not
in the Source_Files and Source_File_List attributes.
The files from --src-subdirs directory should be virtually added into
Interfaces/Library_Interfaces if this attributes is defined because gnatcov
can refer generated sources in the library project from the generated files
of the root standard project.

Change-Id: I0e2beac724b774ecff98495c2e00ec7f40cf912e