Fix the case when config pragma files are removed from project.
T927-001

Detect that the D line file referenced the config pragma file excluding
the case when it is a preprocessor data file. In both cases it is an
absolute pathname, but only in case of a preprocessor data file it is
also referenced from an A line.
If config pragma file is found in D line and not found in the project
then recompile the source code.
Restore config pragma file compilation check for remote compilation,
otherwise there will be spurious recompilation.

Change-Id: I84001f6b1c54182151d63e8eb257537a25dd9958