Fix dependency file conversion
TB21-003

GPRBUILD converts C dependency files to add timestamp for each dependency.
Some characters in dependency file is escaped by GCC compiler. GPRBUILD
normalizes escaped filenames before write it into the new rewritten
dependency file. If escaped filename normalized, it can became name of
different unexistent file.

* gpr/src/gpr-util.adb (Need_To_Compile.Process_Makefile_Deps):
Unescape filename before normalize.

* src/gprbuild-compile.adb (Compilation_Phase.Phase_2_Makefile):
Do not normalize filename before write it into the recreated dependency
file. This Normalize_Filename call wrong because it is normalising filenames
with possible escaped symbols.This normalize is not necessary because if
filename is relative, we are at object directory anyway. When we read
dependency file in Need_To_Compile then we need to normalize it relatevely
object directory because the current directory could be not of the project
object_dir.

Change-Id: I198780e575a0d4dc4cc3bb1af4f013e0baaef836