Remove ALI file copying for gnatprove
U113-006

gnatprove calls gprbuild twice, and this copying was only intended
for the first invocation. This copying being executed in the second
invocation as well is problematic for gnatprove as those ALI files
are read at the same time. We remove this copying of ALI files from
gprbuild, and add it to gnatprove instead, which can do it at a
moment where there are no other readers.

* gprbuild-compile.adb
(Wait_For_Available_Slot): remove copying code which was enabled for
gnatprove only

Change-Id: Icd2bfee421a6236d12a434a0b0696b478d4fdb52
(cherry picked from commit 0b5820e6062988eaa68710feec3ee26519672369)